Here the "like" heart icon has been implemented.
I created a new component called like.jsx.

I converted this from a class to a function to make it a stateless function, so I can get rid of using "this.props" to just "props", as long as "props" is passed into the Like() arrow function.

The heart icon is taken from fontawesome, one is empty and the other is filled, with only "-o" at the end of the string the differs so I was able to dynamically toggle this with boolean onClick events.