Implemented Pagination with working page buttons
Created a new pagination component to render pages depending on itemsCount and pageSize from "props" / or from the movies.jsx component.

In pagination we return an array of pages as a <ul> at the bottom of our movies component in the website.

So far they do render correctly given the a finite number for pageSize, and that the page number won't render if there's not enough items to create a second page.