Created a new JS script called paginate.js to slice an array, take elements, and render them on separate pages.
used the lodash library to correct slice an array, take elements from it depending on the pageSize, and the .value() function returns a new array.

In the movies component I created a brand new redefintion of "movies" to paginate it for each page and the old "movies" variable fro mthe state was renamed to "allMovies" to be used to paginate into "movies"

The pagination function is working.