Created a skeleton on creating two columns; one for features list of movies, the other showing category lists.
Made a new component for listing groups, this will be implemented next commit.

In movies.jsx, now we want to get both movies and genres from fakeMovieServices.js and fakeGenreService.js (Aka. our fake client-sided database).

So I emptied out the movies in the state to be an empty array and also with an empty genre array. I recalled what I learned with Mounting and used componentDidMount(); to initiliaze the movies and genres arrays.

Then in the render() method I created two columns and moved everything to the right column, as for the left column, that will be where our new ListGroup will be.