The left genre column has been rendered
New properties added in movies component to get the name and id of each genre and dynamically added to render in listGroup's render method.

The next syntax that I wasn't aware of was changing

item.name where name is the name is the genre.

To item[textProperty] where "textProperty" is just a string that is "name"