Implemented genre selection highlighting
We're finally using the handleGenreSelect method in movies.jsx, to set the state to create a variable called "selectedGenre" from "genre". Then we made a new prop in <ListGroup> called "selectedItem" to be used in listGroup.jsx.

So in listGroup.jsx, instead of getting the textProperty and valueProperty props from movies.jsx, we can define that here by using "defaultProps".

Then I've created an onClick arrow function event to use the onItemSelect function to pass the item selected, and rendered the highlight if the "item equals the "selectedItem" to active, otherwise it won't have the "active" in its className.