Implemented the Filtering for all movies Genres
Clicking on a genre button to the left should now filter the list by genre.
I've created a new array called "filtered" to filter the allMovies array to the selected genre. I used the length of this filtered array correctly display pagination and correcting the "showing n movies in the database" message at the top.