Implemented Sorting
We take the already filtered array that was have to recreate a new array called "sorted" using the lodash function "orderBy" which takes in 3 parameters (Array, sortPath, Order), the last two parameters are strings.

The handleSort function has been implemented by cloning the sortColumn and then checking if its path is the same then we could check it again if the order is the same, if yes then order becomes 'desc' for descending, else 'asc', which is ascending.